requirements.txt
fastapi==0.79.0
uvicorn==0.18.2
httpx==0.23.0
pytest-asyncio==0.19.0
